Update xsection.lym
Included code, that reduced the negative rounding radius if the rounding would lead to disappearing structures. The same problem for an older version of XSection is discussed here: https://www.klayout.de/forum/discussion/838/xsection-negative-rounding-radius-leads-to-missing-structures#latest

A further optimization would be to only reduce the rounding radius for the affected edge instead of all present edges, but I an not good enough at ruby for that.