WIP: rework of the convolution mechanism
The new concept is to use the minkowsky sum with post-convolution sizing, but account for
narrow valleys and hills: this is done by stretching these locations before convolution,
then apply the sizing and finally remove the stretch. This way, the walls of the valleys
can be covered without merging when they touch.